requirements/base.txt
Click==7.0
pyyaml==5.1.2
boto3==1.9.231
paramiko==2.6.0